+++ src/main/java/com/biezbardis/thoughtyusers/service/AuthenticationServiceImpl.java
@@ -1,0 +1,59 @@
+package com.biezbardis.thoughtyusers.service;
+
+import com.biezbardis.thoughtyusers.dto.AuthenticationRequest;
+import com.biezbardis.thoughtyusers.dto.RegisterRequest;
+import com.biezbardis.thoughtyusers.entity.Role;
+import com.biezbardis.thoughtyusers.entity.User;
+import com.biezbardis.thoughtyusers.exceptions.UnauthorizedException;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class AuthenticationServiceImpl implements AuthenticationService {
+    private final AuthenticationManager authenticationManager;
+    private final JwtService jwtService;
+    private final LoginAttemptService loginAttemptService;
+    private final PasswordEncoder passwordEncoder;
+    private final UserService userService;
+
+    @Override
+    public String register(RegisterRequest request) {
+
+        var user = User.builder()
+                .username(request.getUsername())
+                .email(request.getEmail())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .role(Role.ROLE_USER)
+                .build();
+
+        userService.create(user);
+
+        return jwtService.generateAccessToken(user.getUsername());
+    }
+
+    @Override
+    public String login(AuthenticationRequest request) {
+        Authentication auth;
+        try {
+            auth = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(
+                    request.getUsername(),
+                    request.getPassword()
+            ));
+            loginAttemptService.loginSucceeded(request.getUsername());
+        } catch (AuthenticationException e) {
+            log.info("Authentication failed for user {}", request.getUsername());
+            loginAttemptService.loginFailed(request.getUsername());
+            throw new UnauthorizedException("Invalid credentials", e);
+        }
+
+        return jwtService.generateAccessToken(auth.getName());
+    }
+}
